+++ src/tests/holiday_selector.rs
@@ -16,14 +16,15 @@ fn holidays() -> Result<(), Error> {
         schedule! { 10,00 => Open => 12,00 }
     );
 
-    // The 14th of July is a holiday in the US
+    // Independence Day is a federal holiday. If July 4 is a Saturday, it is
+    // observed on Friday, July 3.
     assert_eq!(
-        schedule_at!("2020:10:00-12:00; PH off", "2020-07-04", "US"),
+        schedule_at!("2020:10:00-12:00; PH off", "2020-07-03", "US"),
         schedule! { 00,00 => Closed => 24,00 }
     );
 
     assert_eq!(
-        schedule_at!("2020:10:00-12:00; PH off", "2020-07-04", "FR"),
+        schedule_at!("2020:10:00-12:00; PH off", "2020-07-04", "US"),
         schedule! { 10,00 => Open => 12,00 }
     );